Package expects `README.adoc`
The `.spec` notes a `%doc` of `README.adoc`, this converts the `.md`
back to `.adoc` to meet this requirement.

Move the download code into a script to promote local builds.

This fixes the version so the RPM includes the spire version.

It also tidies the `README` up, and removes incorrect information.

This also refreshes the `Makefile`, removing `BUILD_METADATA`

Now a shellcheck is ran for PRs, this PR includes shellcheck fixes.

Lastly this adds licenses to the old `conf/` files that pre-date the
GitHub transition.